+++ CHANGELOG.md
@@ -5,7 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Fix Warning on default commissionRules prop type for LoanSimulator.
-- Fix: Disable React on Rails console traces in dev style guide.
+- Feature: Handle multi-level options in the Select.
+- Feature: Select can now be clearable, searchable, and multi.
+- Feature: Add new `k-VerticalGrid` component.
+- Feature: Add new `k-FormActions` component.
+- Fix: disable React on Rails console traces in dev style guide.
 
 ## [v.4.4.0] - 2016-12-27